requirements.txt
Django==4.2.1
django-cors-headers==3.11.0
django-crispy-forms==1.12.0
django-filter==21.1
django-jquery==3.1.0
django-model-utils==4.1.1
django-taggit==3.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
